assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools",
    # 4110 only:
    # "cython",
    # "numpy==1.21.*",
]
build-backend = "setuptools.build_meta"

[project]
version = "0.1.0"
requires-python = ">=3.7"
license = {text = "MIT License"}
name = "Instapy"
description = "A tool to filter images"
readme = "README.md"
dependencies = [
	"numpy==1.21.*",
	"numba",
	"pillow",
]


[project.scripts]
instapy = "instapy.cli:main"
